fix tests and circular imports, update dependencies
