+++ .secrets.baseline
@@ -4,7 +4,7 @@
     "files": null,
     "lines": null
   },
-  "generated_at": "2020-08-28T02:45:35Z",
+  "generated_at": "2020-08-28T04:04:46Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -20,8 +20,17 @@
       "name": "BasicAuthDetector"
     },
     {
+      "name": "CloudantDetector"
+    },
+    {
       "hex_limit": 3,
       "name": "HexHighEntropyString"
+    },
+    {
+      "name": "IbmCloudIamDetector"
+    },
+    {
+      "name": "IbmCosHmacDetector"
     },
     {
       "name": "JwtTokenDetector"
@@ -44,9 +53,21 @@
     },
     {
       "name": "StripeDetector"
+    },
+    {
+      "name": "TwilioKeyDetector"
     }
   ],
-  "results": {},
+  "results": {
+    "Pipfile.lock": [
+      {
+        "hashed_secret": "2a54441c42f9c61ccd3ab33bc92f6445a3fe0282",
+        "is_verified": false,
+        "line_number": 4,
+        "type": "Hex High Entropy String"
+      }
+    ]
+  },
   "version": "0.14.2",
   "word_list": {
     "file": null,
