+++ Pipfile
@@ -4,25 +4,25 @@ url = "https://pypi.org/simple"
 verify_ssl = true
 
 [dev-packages]
-pytest = "~=5.2.1"
-pytest-cov = "~=2.8.1"
+pytest = "*"
+pytest-cov = "*"
 pre-commit = "*"
 
 [packages]
-boto3 = "~=1.10.1"
-click = "~=7.0"
-geojson = "~=2.5.0"
-parallelpipe = "~=0.2.6"
-psutil = "~=5.6.7"
-psycopg2 = "~=2.8.4"
-pydantic = "~=1.6.1"
-pyproj = "~=2.4.0"
-retrying = "~=1.3.3"
-PyYAML = "~=5.1.2"
-Shapely = "~=1.6.4.post2"
-SQLAlchemy = "~=1.3.18"
+boto3 = "*"
+click = "*"
+geojson = "*"
+parallelpipe = "*"
+psutil = "*"
+psycopg2 = "*"
+pydantic = "*"
+pyproj = "*"
+retrying = "*"
+PyYAML = "*"
+Shapely = "*"
+SQLAlchemy = "*"
 rasterio = "*"
 google-cloud-storage = "*"
 
 [requires]
-python_version = "3.7"
+python_version = "3.8"
