+++ gfw_pixetl/utils/__init__.py
@@ -1,1 +1,11 @@
-from gfw_pixetl.utils.utils import *
+from gfw_pixetl.utils.utils import (
+    available_memory_per_process,
+    get_bucket,
+    get_module_logger,
+    get_workers,
+    set_available_memory,
+    set_workers,
+    snapped_window,
+    verify_version_pattern,
+    world_bounds,
+)
